requirements.txt
numpy==1.16.6
scipy==1.4.1
pillow
opencv-python
scikit-image==0.16.2
lz4
tqdm
requests<2.19.0,>=2.18.4
uiautomator2==2.9.2
retrying
mxnet==1.6.0
cnocr==1.2.2
gooey
colored
# python-Levenshtein
# win10toast-persist
notify-py
jellyfish
toml
pymysql
ruia==0.8.4